uguhuh

+++ excelBuilder.groovy
@@ -28,7 +28,7 @@ sheet.setColumnWidth (5, 11000)
 sheet.setColumnWidth (6, 7500)
 
 // Erstelle eine Kopfzeile
-XSSFRow headerRow = sheet.createRow((short) 0)
+XSSFRow headerRow = sheet.createRow(0)
 headerRow.setHeightInPoints((int) 35)
 headerRow.createCell(0).setCellValue("Kapitän")
 headerRow.createCell(1).setCellValue("Vorname Nachname")
